Major overhaul to delete endpoints. And start building ram checking

+++ job_service/app/crud.py
@@ -4,6 +4,7 @@ from dateutil import parser
 import datetime
 import uuid
 import pytz
+import executor
 
 import models, schemas
 
@@ -19,12 +20,27 @@ def get_solver_instances(db: Session, job_id: str, user_id: str):
     solvers = db.query(models.Job).filter(models.Job.user_id == user_id).filter(models.Job.id == job_id).first().solver_instances
     return list(solvers)
 
-def delete_job(db: Session, job_id: str, user_id: str):
+def get_user_from_job(db: Session, job_id: str):
+    job = db.query(models.Job).filter(models.Job.id == job_id).first()
+    return job.user_id
+
+def delete_job(db: Session, job_id: str):
     #TODO: Check that user owns this JOB
-    job = db.query(models.Job).filter(models.Job.id == job_id).filter(models.Job.user_id == user_id).first()
-    db.delete(job)
-    db.commit()
-    return {"success"}
+    job = db.query(models.Job).filter(models.Job.id == job_id).first()
+    if job:
+      db.delete(job)
+      db.commit()
+
+      print("attempting to stop job")
+      for solver in job.solver_instances:
+        try:
+          executor.stop_job(solver.id, "default")
+          print(f"successfully stopped solver: {solver.id}")
+        except:
+          print(f"failed to stopped solver: {solver.id} (it was already stopped)")
+      return {"success"}
+    else:
+      return {"failure"}
 
 def delete_all_jobs(db: Session, user_id: str):
     # TODO: Test this
@@ -40,8 +56,14 @@ def stop_solver(db: Session, job_id: str, solver_id: str, user_id: str):
         if str(solver.id) == str(solver_id):
           db.delete(solver)
           db.commit()
-          return {"success"}
-    return {"failed to stop solver"}
+    print("attempting to stop solver")
+    try:
+      executor.stop_job(solver_id, "default")
+      print("successfully stopped solver")
+      return {"success"}
+    except:
+      print("failed to stopped solver (it was already stopped)")
+      return {"success"}
 
 def solvers_left(db, job_id: str):
     job = db.query(models.Job).filter(models.Job.id == job_id).first()
@@ -104,3 +126,25 @@ def num_vcpus_in_use(db, user_id: str):
           vcpu_sum += solver.vcpus
           print(vcpu_sum)
     return vcpu_sum
+
+def ram_in_use(db, user_id: str):
+    jobs = db.query(models.Job).filter(models.Job.user_id == user_id)
+    ram_sum = 0
+    for job in jobs:
+      for solver in job.solver_instances:
+        if solver.status == "running" and solver.time_created + datetime.timedelta(hours=1,seconds=solver.timeout) > utc.localize(datetime.datetime.now()):
+          ram_sum += solver.ram
+          print(ram_sum)
+    return ram_sum
+
+def set_pod_name(db, job_id: str, solver_id: str, pod_name: str):
+    print(f"job_id: {job_id}")
+    print(f"solver_id: {solver_id}")
+    print(f"pod_name: {pod_name}")
+    solvers = db.query(models.Job).filter(models.Job.id == job_id).first().solver_instances
+    for solver in solvers:
+        if str(solver.id) == str(solver_id):
+          solver.pod_name = pod_name
+          db.commit()
+          return {"success"}
+    return {"error"}
